+++ crates/acts/src/package/core/parallel.rs
@@ -19,6 +19,7 @@ struct ParallelPackageParams {
     acts: Vec<Act>,
 }
 
+#[async_trait::async_trait]
 impl ActPackage for ParallelPackage {
     fn definition() -> ActPackageDefinition {
         ActPackageDefinition {
@@ -73,7 +74,7 @@ impl ActPackage for ParallelPackage {
         Ok(Self)
     }
 
-    fn execute(&self, ctx: &Context, params: &serde_json::Value) -> Result<Option<Vars>> {
+    async fn execute(&self, ctx: &Context, params: &serde_json::Value) -> Result<Option<Vars>> {
         let params =
             serde_json::from_value::<ParallelPackageParams>(params.clone()).map_err(|e| {
                 ActError::Package(format!(