+++ crates/acts/src/scheduler/process/task/act.rs
@@ -57,7 +57,7 @@ impl ActTask for Act {
                         package.id
                     )))?;
                 let package = (register.create)(ctx.runtime.config())?;
-                if let Some(vars) = package.execute(ctx, &ctx.task().params())? {
+                if let Some(vars) = crate::utils::sync::block_on(package.execute(ctx, &ctx.task().params()))? {
                     task.update_data(&vars);
                 };
             }